Use costume loading from importer in public costume and backdrop loader
